Pulled logic out of OrdersBlockController
